# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"OROReasoningEffort": {
				"type": "string",
				"enum": [
					"minimal",
					"low",
					"medium",
					"high",
					"xhigh"
				]
			},
			"GAISystemPromptMode": {
				"enum": [
					"LOCAL",
					"CONTEXT"
				],
				"type": "string"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "jaipuroxy",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/zai/thinky/chat": {
			"post": {
				"operationId": "ZaiProxyChat",
				"responses": {
					"200": {
						"description": "JAI compatible response from Z.AI.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Error"
					}
				},
				"description": "Uses the Z.AI chat API endpoint for completion, enables reasoning.",
				"security": [],
				"parameters": [
					{
						"description": "Mandatory. Needs to be 'Bearer <Z.AI_API_KEY>', automatically set by JAI if you put in your Z.AI API key in the proxy config.",
						"in": "header",
						"name": "authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Optional. For running locally: If available, reasoning will be logged to console.",
						"in": "query",
						"name": "logReasoning",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"description": "The JAI request body.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "The JAI request body."
							}
						}
					}
				}
			}
		},
		"/api/zai/thinky/coding": {
			"post": {
				"operationId": "ZaiProxyCoding",
				"responses": {
					"200": {
						"description": "JAI compatible response from Z.AI.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Error"
					}
				},
				"description": "Uses the Z.AI coding API endpoint for completion, enables reasoning.",
				"security": [],
				"parameters": [
					{
						"description": "Mandatory. Needs to be 'Bearer <Z.AI_API_KEY>', automatically set by JAI if you put in your Z.AI API key in the proxy config.",
						"in": "header",
						"name": "authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Optional. For running locally: If available, reasoning will be logged to console.",
						"in": "query",
						"name": "logReasoning",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"description": "The JAI request body.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "The JAI request body."
							}
						}
					}
				}
			}
		},
		"/api/oro/thinky": {
			"post": {
				"operationId": "OroProxy",
				"responses": {
					"200": {
						"description": "JAI compatible response from OpenRouter.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Error"
					}
				},
				"description": "Uses OpenRouter endpoint for completion, enables reasoning.",
				"security": [],
				"parameters": [
					{
						"description": "Optional. Enables use of a preset configured in your OpenRouter account, e.g. to set the system prompt or customise providers. Example:",
						"in": "query",
						"name": "preset",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "reasoningEffort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OROReasoningEffort"
						}
					},
					{
						"description": "Optional. For running locally: If available, reasoning will be logged to console.",
						"in": "query",
						"name": "logReasoning",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"description": "Needs to be 'Bearer <OpenRouter_API_KEY>', automatically set by JAI if you put in your OpenRouter API key in the proxy config.  If empty/blank, attempts to use .env variables for ORO_API_KEY.",
						"in": "header",
						"name": "authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The JAI request body.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "The JAI request body."
							}
						}
					}
				}
			}
		},
		"/api/gai/proxy": {
			"post": {
				"operationId": "GaiProxy",
				"responses": {
					"200": {
						"description": "JAI compatible response from GAI.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Error"
					}
				},
				"description": "Translates a JAI request into a Google AI request, sends it to GAI, then translates the response back to JAI format.",
				"security": [],
				"parameters": [
					{
						"description": "Needs to be 'Bearer <GoogleAI_API_KEY>', automatically set by JAI if you put in your Google AI Studio API key in the proxy config. If empty/blank, attempts to use .env variables for GAI_API_KEY_PAID (pro models) and GAI_API_KEY_FREE (other models).",
						"in": "header",
						"name": "authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Optional. Sets \"system instructions\" which the model will adhere more strictly to. LOCAL loads the system prompt in the server file system, i.e. systemprompt.md - use if you are running it locally, or you want to use the server's \"default\" system prompt. CONTEXT reads whatever contents put between <systemprompt>...</systemprompt> in the first message of the conversation and uses that as the system prompt. If not specified, no system prompt is set.",
						"in": "query",
						"name": "systemPromptMode",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/GAISystemPromptMode"
						}
					},
					{
						"description": "Optional. Lets you set a determined number of tokens the model may use for reasoning, if you are using a model that is capable of reasoning. Setting this to a positive integer can force gemini-2.5-flash into reasoning mode. For flash: 0 to 24576, for pro: 128 to 32768. -1 lets the model decide.",
						"in": "query",
						"name": "reasoningEffort",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Optional. For running locally: If available, reasoning will be logged to console.",
						"in": "query",
						"name": "logReasoning",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"description": "Mandatory. The JAI request body.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "Mandatory. The JAI request body."
							}
						}
					}
				}
			}
		},
		"/api/aux/snip": {
			"post": {
				"operationId": "Snip",
				"responses": {
					"200": {
						"description": "JAI compatible response.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Internal Error"
					}
				},
				"description": "Splits up bot's most recent response at <SNIP> tag and returns the second part as a \"bot\" response again.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The JAI request body.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "The JAI request body."
							}
						}
					}
				}
			}
		},
		"/api/aux/log": {
			"post": {
				"operationId": "Log",
				"responses": {
					"200": {
						"description": "JAI compatible response.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"418": {
						"description": "Context Logged"
					},
					"500": {
						"description": "Internal Error"
					}
				},
				"description": "Logs the full context currently sent to the LLM.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The JAI request body.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "The JAI request body."
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8089",
			"description": "localhost (default port)"
		},
		{
			"url": "https://jaipuroxy.onrender.com",
			"description": "OnRender Deployment"
		}
	]
}